Refactor ProfileEdit and Playlist into smaller composable
