+++ app/src/main/java/com/example/buibinhminh/data/Song.kt
@@ -1,0 +1,8 @@
+package com.example.buibinhminh.data
+
+data class Song(
+    val id: Int,
+    val songName: String,
+    val authorName: String,
+    val duration: String
+)
