+++ app/src/main/java/com/example/buibinhminh/ui/musicApp/SongGridItem.kt
@@ -1,0 +1,74 @@
+package com.example.buibinhminh.ui.musicApp
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.buibinhminh.R
+import com.example.buibinhminh.data.Song
+
+@Composable
+fun SongGridItem(
+    song: Song,
+    onDeleteClick: (Song) -> Unit)
+{
+    Column(
+        modifier = Modifier
+            .padding(16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Box {
+            Image(
+                painter = painterResource(id = R.drawable.song),
+                contentDescription = "Song",
+                modifier = Modifier
+                    .padding(8.dp)
+                    .clip(RoundedCornerShape(10.dp))
+            )
+            SongOptionButton(
+                song = song,
+                onDeleteClick = onDeleteClick,
+                modifier = Modifier
+                    .align(Alignment.TopEnd)
+                    .padding(top = 16.dp, end = 16.dp)
+                    .size(36.dp)
+                    .clip(CircleShape)
+                    .background(Color.Black.copy(alpha = 0.5f))
+            )
+        }
+
+        Text(
+            text = song.songName,
+            fontSize = 20.sp,
+            fontWeight = FontWeight.Bold,
+            color = Color.White,
+            modifier = Modifier.padding(bottom = 4.dp)
+        )
+        Text(
+            text = song.authorName,
+            fontSize = 18.sp,
+            fontWeight = FontWeight.SemiBold,
+            color = Color.Gray,
+            modifier = Modifier.padding(bottom = 8.dp)
+        )
+        Text(
+            text = song.duration,
+            fontSize = 17.sp,
+            color = Color.White,
+        )
+    }
+}
