+++ app/src/main/java/com/example/buibinhminh/ui/musicApp/SongListItem.kt
@@ -1,0 +1,82 @@
+package com.example.buibinhminh.ui.musicApp
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.buibinhminh.R
+import com.example.buibinhminh.data.Song
+
+@Composable
+fun SongListItem(
+    song: Song,
+    onDeleteClick: (Song) -> Unit
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(80.dp),
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Row {
+            Image(
+                painter = painterResource(id = R.drawable.song),
+                contentDescription = "Song",
+                modifier = Modifier
+                    .padding(8.dp)
+                    .clip(RoundedCornerShape(10.dp))
+            )
+            Column(
+                modifier = Modifier
+                    .fillMaxHeight()
+                    .padding(8.dp, 18.dp),
+                verticalArrangement = Arrangement.SpaceBetween
+            ) {
+                Text(
+                    text = song.songName,
+                    fontSize = 16.sp,
+                    fontWeight = FontWeight.Bold,
+                    color = Color.White
+                )
+                Text(
+                    text = song.authorName,
+                    fontSize = 14.sp,
+                    fontWeight = FontWeight.SemiBold,
+                    color = Color.Gray
+                )
+            }
+        }
+        Row(
+            modifier = Modifier
+                .fillMaxHeight()
+                .padding(8.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = song.duration,
+                fontSize = 17.sp,
+                color = Color.White
+            )
+            SongOptionButton(
+                song = song,
+                onDeleteClick = onDeleteClick,
+                modifier = Modifier.padding(start = 8.dp)
+            )
+        }
+    }
+}
